updates based on official and updated package
