+++ custom_components/sia/manifest.json
@@ -4,7 +4,7 @@
   "config_flow": true,
   "documentation": "https://www.home-assistant.io/integrations/sia",
   "requirements": [
-    "pysiaalarm==2.0.7"
+    "pysiaalarm==2.0.8"
   ],
   "codeowners": [
     "@eavanvalkenburg"
